+++ src/Sylius/Bundle/CartBundle/Controller/CartItemController.php
@@ -13,8 +13,8 @@ namespace Sylius\Bundle\CartBundle\Controller;
 
 use Sylius\Bundle\CartBundle\Event\CartItemEvent;
 use Sylius\Bundle\CartBundle\Event\FlashEvent;
-use Sylius\Bundle\CartBundle\SyliusCartEvents;
 use Sylius\Component\Cart\Resolver\ItemResolvingException;
+use Sylius\Component\Cart\SyliusCartEvents;
 use Symfony\Component\EventDispatcher\GenericEvent;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
