+++ src/Sylius/Bundle/CartBundle/EventListener/FlashListener.php
@@ -12,7 +12,7 @@
 namespace Sylius\Bundle\CartBundle\EventListener;
 
 use Sylius\Bundle\CartBundle\Event\FlashEvent;
-use Sylius\Bundle\CartBundle\SyliusCartEvents;
+use Sylius\Component\Cart\SyliusCartEvents;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 use Symfony\Component\HttpFoundation\Session\SessionInterface;
 use Symfony\Component\Translation\TranslatorInterface;
