+++ src/Sylius/Bundle/CoreBundle/Checkout/Step/FinalizeStep.php
@@ -11,10 +11,10 @@
 
 namespace Sylius\Bundle\CoreBundle\Checkout\Step;
 
-use Sylius\Bundle\CoreBundle\Checkout\SyliusCheckoutEvents;
 use Sylius\Bundle\FlowBundle\Process\Context\ProcessContextInterface;
-use Sylius\Bundle\OrderBundle\SyliusOrderEvents;
 use Sylius\Component\Core\Model\OrderInterface;
+use Sylius\Component\Core\SyliusCheckoutEvents;
+use Sylius\Component\Core\SyliusOrderEvents;
 
 /**
  * Final checkout step.
