+++ src/Sylius/Bundle/CoreBundle/spec/Sylius/Bundle/CoreBundle/Checkout/Step/PurchaseStepSpec.php
@@ -19,12 +19,12 @@ use Payum\Core\Security\TokenInterface;
 use PhpSpec\ObjectBehavior;
 use Prophecy\Argument;
 use spec\Sylius\Bundle\CoreBundle\Fixture\RequestStack;
-use Sylius\Bundle\CoreBundle\Checkout\SyliusCheckoutEvents;
 use Sylius\Bundle\FlowBundle\Process\Context\ProcessContextInterface;
-use Sylius\Bundle\PaymentBundle\SyliusPaymentEvents;
 use Sylius\Component\Cart\Provider\CartProviderInterface;
 use Sylius\Component\Core\Model\Order;
+use Sylius\Component\Core\SyliusCheckoutEvents;
 use Sylius\Component\Payment\Model\Payment;
+use Sylius\Component\Payment\SyliusPaymentEvents;
 use Symfony\Bridge\Doctrine\RegistryInterface as DoctrinRegistryInterface;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
