+++ src/Sylius/Bundle/ProductBundle/Resources/config/doctrine/model/Variant.orm.xml
@@ -15,8 +15,7 @@
                   xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
                   xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping
                                       http://doctrine-project.org/schemas/orm/doctrine-mapping.xsd"
-                  xmlns:gedmo="http://gediminasm.org/schemas/orm/doctrine-extensions-mapping"
->
+        >
 
     <mapped-superclass name="Sylius\Component\Product\Model\Variant" table="sylius_product_variant">
         <field name="availableOn" column="available_on" type="datetime" />
