+++ src/Sylius/Bundle/VariationBundle/EventListener/LoadMetadataSubscriber.php
@@ -13,7 +13,6 @@ namespace Sylius\Bundle\VariationBundle\EventListener;
 
 use Doctrine\Common\EventSubscriber;
 use Doctrine\ORM\Event\LoadClassMetadataEventArgs;
-use Doctrine\ORM\Mapping\ClassMetadata;
 use Doctrine\ORM\Mapping\ClassMetadataInfo;
 
 /**
