+++ src/Sylius/Bundle/VariationBundle/Form/Type/OptionValueChoiceType.php
@@ -11,7 +11,6 @@
 
 namespace Sylius\Bundle\VariationBundle\Form\Type;
 
-use Sylius\Component\Variation\Model\OptionInterface;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\Extension\Core\ChoiceList\ObjectChoiceList;
 use Symfony\Component\OptionsResolver\Options;
