+++ src/Sylius/Bundle/VariationBundle/spec/Sylius/Bundle/VariationBundle/Form/Type/VariantTypeSpec.php
@@ -13,7 +13,6 @@ namespace spec\Sylius\Bundle\VariationBundle\Form\Type;
 
 use PhpSpec\ObjectBehavior;
 use Prophecy\Argument;
-use Symfony\Component\Form\FormBuilder;
 use Symfony\Component\OptionsResolver\OptionsResolverInterface;
 
 /**
